Make it harder to optimize the asserts away
In theory, the asserts can't fail, because RAM content juts doesn't
change on its own in the memory model. Therefore, rustc could just
optimize them away and be done with it.

However, we are actually checking that the content doesn't change on its
own, so we make it harder to optimize them away by hint::black_box.